refactor: modularize MechSilhouette.tsx into focused files
Split 1346-line MechSilhouette.tsx into 7 focused modules:
- MechSilhouette.types.ts (15 lines): Type definitions
- MechSilhouette.biped.ts (488 lines): Biped silhouettes (realistic, battlemech, megamek, geometric)
- MechSilhouette.quad.ts (145 lines): Quad mech silhouette
- MechSilhouette.tripod.ts (201 lines): Tripod mech silhouette
- MechSilhouette.fighter.ts (78 lines): LAM fighter silhouette
- MechSilhouette.labels.ts (46 lines): Location label mappings
- MechSilhouette.utils.ts (42 lines): Utility functions

Main file (MechSilhouette.tsx) is now a 35-line aggregator.
All 5301 tests passing, backward compatibility maintained.
